feature/unittests (#134)
* bugfix: bitmap clear method

* feature/unittests

* add license

* support cmake-maven

* support cmake-maven

* unittest support cmake-maven

* unittest support cmake-maven

* fix for macos compilation

* fix for macos compilation

* fix for macos compilation

* fix for macos compilation

* fix unittest dependency

* fix macos compilation

* fix win compilation

* fix win compilation

* fix win compilation

* test

* test

* test

* fix windows mingw compilation

* fix mingw compilation

* test

* fix MinGW compilation

* fix MinGW compilation

* fix MinGW compilation

* fix MinGW compilation

* test compress

* test/encoding

* test/write

* testcompress

* testencoding

* testcommon

* testutils

* testfile

* fix MinGW compilation

* support cross platform compilation

* support mingw macos ubuntu compilation

* fix code formatting issues

* fix newly introduced encoding issue(bitpack)

---------

+++ cpp/src/common/allocator/byte_stream.h
@@ -220,7 +220,7 @@ FORCE_INLINE double bytes_to_double(uint8_t bytes[8]) {
 
     // follow jdk implementation style.
     union {
-        long l;
+        long long l;
         double d;
     } u;
